Bug 1967047: Console overview operators shown upgrading when still waiting on approval
