+++ frontend/packages/operator-lifecycle-manager/src/components/clusterserviceversion.tsx
@@ -98,7 +98,11 @@ import {
   PackageManifestKind,
   SubscriptionKind,
 } from '../types';
-import { getClusterServiceVersionPlugins, isPluginEnabled } from '../utils';
+import {
+  getClusterServiceVersionPlugins,
+  isPluginEnabled,
+  upgradeRequiresApproval,
+} from '../utils';
 import { consolePluginModal } from './modals/console-plugin-modal';
 import { createUninstallOperatorModal } from './modals/uninstall-operator-modal';
 import { ProvidedAPIsPage, ProvidedAPIPage } from './operand';
@@ -107,7 +111,6 @@ import { CreateInitializationResourceButton } from './operator-install-page';
 import {
   SubscriptionDetails,
   catalogSourceForSubscription,
-  upgradeRequiresApproval,
   UpgradeApprovalLink,
 } from './subscription';
 import { ClusterServiceVersionLogo, referenceForProvidedAPI, providedAPIsForCSV } from './index';
